.ipynb_checkpoints/requirements.txt
#実際に、どういうライブラリを使っているのかを知らせないと、うまく動かない。
#標準ライブラリ出ないものをかく。
#==でversionを✖かける・
streamlit

Pillow
